fix: implement multi-noun interaction for lighting matches with torch

+++ UnitTests/TestParser.cs
@@ -596,6 +596,16 @@ public class TestParser : IntentParser
                 Verb = "light",
                 OriginalInput = "light candles with torch"
             });
+            
+        if (input == "light matches with torch")
+            return Task.FromResult<IntentBase>(new MultiNounIntent
+            {
+                NounOne = "matches",
+                NounTwo = "torch",
+                Preposition = "with",
+                Verb = "light",
+                OriginalInput = "light matches with torch"
+            });
 
         if (input == "kill the cyclops with the sword")
             return Task.FromResult<IntentBase>(new MultiNounIntent
